derp

+++ katt/friends/include/cpp/friends.h
@@ -2,4 +2,4 @@ void init(int N, int L, int P[]);
 
 void jump(int A, int B);
 
-int score();
+long long score();
